rename hardware discovery module
